requirements.txt
asgiref==3.4.1
celery==4.4.2
celery-rmq==0.0.2
coverage==6.4
Django==4.0.1
django-debug-toolbar==3.4.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-smtp-ssl==1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
factory-boy==3.2.1
Faker==13.12.0
model-bakery==1.5.0
PyJWT==2.4.0
pytest==7.1.2
pytest-celery==0.0.0
pytest-django==4.5.2
pytest-factoryboy==2.3.1
pytz==2022.1
sqlparse==0.4.2